+++ svg/svg_test.go
@@ -95,11 +95,3 @@ func BenchmarkGetAttributes(b *testing.B) {
 		getAttributes(&attrTokenBuffer, tb, svg.X, svg.Y, svg.Width, svg.Height, svg.Rx, svg.Ry)
 	}
 }
-
-func BenchmarkShortenPathData(b *testing.B) {
-	pathDataBuffer := &PathData{}
-	r := []byte("M8.64,223.948c0,0,143.468,3.431,185.777-181.808c2.673-11.702-1.23-20.154,1.316-33.146h16.287c0,0-3.14,17.248,1.095,30.848c21.392,68.692-4.179,242.343-204.227,196.59L8.64,223.948z")
-	for i := 0; i < b.N; i++ {
-		ShortenPathData(r, pathDataBuffer)
-	}
-}
